Best BLASTP Hits

Best BLASTP Hits

blastp(version = NULL, dir = ".")

Arguments

version
WormBase version. If NULL, defaults to current production release version available on the WormBase website. Legacy versions can be specified as a string (e.g. "WS262").

dir
Output directory.

Value

tibble grouped by wormpep.

See also

Other FTP File Functions: description, geneIDs, geneOtherIDs, oligos, orthologs, peptides, rnaiPhenotypes

Examples

blastp() %>% glimpse()

#> Downloading c_elegans.PRJNA13758.WS262.best_blastp_hits.txt.gz

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   X3 = col_double(),
#>   X5 = col_double(),
#>   X7 = col_double(),
#>   X9 = col_double(),
#>   X11 = col_double(),
#>   X13 = col_double(),
#>   X15 = col_double(),
#>   X17 = col_double(),
#>   X19 = col_double(),
#>   X21 = col_double(),
#>   X23 = col_double(),
#>   X25 = col_double(),
#>   X27 = col_double()
#> )

#> See spec(...) for full column specifications.

#> Observations: 14,533
#> Variables: 3
#> $ wormpep <chr> "CE00004", "CE00015", "CE00034", "CE00063", "CE00072", "CE0...
#> $ peptide <chr> "ENSP00000308897", "ENSP00000444192", "ENSP00000384886", "E...
#> $ eValue  <dbl> 1e-46, 0e+00, 4e-19, 8e-08, 1e-20, 9e-66, 7e-07, 2e-25, 9e-...