Other gene identifiers

Other gene identifiers

geneOtherIDs(version = getOption(x = "wormbase.version", default = NULL))

Arguments

version
character(1) or NULL. WormBase version. If NULL (recommended), defaults to current production release version available on the WormBase website. Legacy versions can be specified as a character string (e.g. "WS262").

Value

DataFrame.

Note

As of WS269 release, some non-N2 gene IDs are included in the flat files available on the WormBase FTP server. These annotations are removed from the return here, using grep matching to return only WBGene entries.

Updated 2019-08-28.

Examples

## WormBase FTP server must be accessible.
tryCatch(
    expr = geneOtherIDs(),
    error = function(e) e
)

#> Transmitting files from <ftp.wormbase.org>.

#> ℹ Files matching pattern: c_elegans.PRJNA13758.WS277.geneOtherIDs.txt.gz

#> → Downloading c_elegans.PRJNA13758.WS277.geneOtherIDs.txt.gz.

#> → Importing c_elegans.PRJNA13758.WS277.geneOtherIDs.txt.gz at /private/var/folders/l1/8y8sjzmn15v49jgrqglghcfr0000gn/T/RtmpRU0dkX using readr::`read_lines()`.

#> DataFrame with 52086 rows and 2 columns
#>               geneID     geneOtherIDs
#>          <character>           <list>
#> 1     WBGene00000001 Y110A7A.10,aap-1
#> 2     WBGene00000002    F27C8.1,aat-1
#> 3     WBGene00000003    F07C3.7,aat-2
#> 4     WBGene00000004    F52H2.2,aat-3
#> 5     WBGene00000005  T13A10.10,aat-4
#> ...              ...              ...
#> 52082 WBGene00306071            mha-1
#> 52083 WBGene00306072            mha-2
#> 52084 WBGene00306073            mha-3
#> 52085 WBGene00306076       Y53C10A.25
#> 52086 WBGene00306078        Y51H7C.42