PCR oligo sequences

PCR oligo sequences

oligos(version = getOption(x = "wormbase.version", default = NULL))

Arguments

version
character(1) or NULL. WormBase version. If NULL (recommended), defaults to current production release version available on the WormBase website. Legacy versions can be specified as a character string (e.g. "WS262").

Value

DataFrame.

Note

Updated 2019-08-28.

Examples

## WormBase FTP server must be accessible.
tryCatch(
    expr = oligos(),
    error = function(e) e
)

#> Transmitting files from 'ftp.wormbase.org'.

#> Files matching pattern:
#> c_elegans.PRJNA13758.WS274.pcr_product2gene.txt.gz

#> Skipped: c_elegans.PRJNA13758.WS274.pcr_product2gene.txt.gz.

#> All files are already downloaded.

#> Decompressing 'c_elegans.PRJNA13758.WS274.pcr_product2gene.txt.gz' in 'tempdir()'.

#> Importing 'c_elegans.PRJNA13758.WS274.pcr_product2gene.txt.gz' using 'base::readLines()'.

#> DataFrame with 18588 rows and 2 columns
#>               geneID
#>          <character>
#> 1     WBGene00000001
#> 2     WBGene00000002
#> 3     WBGene00000003
#> 4     WBGene00000004
#> 5     WBGene00000005
#> ...              ...
#> 18584 WBGene00304820
#> 18585 WBGene00304823
#> 18586 WBGene00304825
#> 18587 WBGene00304826
#> 18588 WBGene00304981
#>                                                                     oligo
#>                                                                    <list>
#> 1                      c("cenix:90-h3", "mv_Y110A7A.10", "sjj_Y110A7A.k")
#> 2                           c("cenix:38-c8", "mv_F27C8.1", "sjj_F27C8.1")
#> 3                           c("cenix:23-c7", "mv_F07C3.7", "sjj_F07C3.7")
#> 4                           c("cenix:65-h1", "mv_F52H2.2", "sjj_F52H2.2")
#> 5     c("cenix:28-e8", "mv_T13A10.10", "sjj_T13A10.10", "sjj2_T13A10.10")
#> ...                                                                   ...
#> 18584                                     c("cenix:114-f4", "sjj_R193.b")
#> 18585       c("cenix:151-g2", "mv_Y7A5A.4", "sjj_Y7A5A.4", "smd_Y7A5A.4")
#> 18586                                                        cenix:110-b6
#> 18587                                                       cenix:190-h10
#> 18588                                                         cenix:33-a4