Orthologs

Orthologs

orthologs(
  version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam()
)

Arguments

version
character(1) or NULL. WormBase version. If NULL (recommended), defaults to current production release version available on the WormBase website. Legacy versions can be specified as a character string (e.g. "WS262").

BPPARAM
bpparamClass. BiocParallel parameter to specify the desired processor configuration.
We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

Value

DataFrame.

Note

Updated 2020-01-26.

Examples

## WormBase FTP server must be accessible.
tryCatch(
    expr = orthologs(),
    error = function(e) e
)

#> Transmitting files from <ftp.wormbase.org>.

#> ℹ Files matching pattern: c_elegans.PRJNA13758.WS276.orthologs.txt.gz

#> → Downloading c_elegans.PRJNA13758.WS276.orthologs.txt.gz.

#> → Importing c_elegans.PRJNA13758.WS276.orthologs.txt.gz using readr::`read_lines()`.

#> → Processing orthologs.

#> DataFrame with 20126 rows and 5 columns
#>               geneID
#>          <character>
#> 1     WBGene00000001
#> 2     WBGene00000002
#> 3     WBGene00000003
#> 4     WBGene00000004
#> 5     WBGene00000005
#> ...              ...
#> 20122 WBGene00306002
#> 20123 WBGene00306003
#> 20124 WBGene00306004
#> 20125 WBGene00306005
#> 20126 WBGene00306013
#>                                                         danioRerio
#>                                                             <list>
#> 1     ENSDARG00000018060,ENSDARG00000034409,ENSDARG00000038524,...
#> 2     ENSDARG00000005894,ENSDARG00000008100,ENSDARG00000020278,...
#> 3     ENSDARG00000005894,ENSDARG00000008100,ENSDARG00000020278,...
#> 4     ENSDARG00000005894,ENSDARG00000008100,ENSDARG00000010816,...
#> 5     ENSDARG00000005894,ENSDARG00000008100,ENSDARG00000010816,...
#> ...                                                            ...
#> 20122                                                             
#> 20123                                                             
#> 20124                                                             
#> 20125                                                             
#> 20126                                                             
#>                        drosophilaMelanogaster
#>                                        <list>
#> 1                                 FBgn0020622
#> 2                                 FBgn0039844
#> 3         FBgn0002778,FBgn0028425,FBgn0039487
#> 4     FBgn0031064,FBgn0032911,FBgn0036493,...
#> 5     FBgn0028425,FBgn0030574,FBgn0031064,...
#> ...                                       ...
#> 20122                                        
#> 20123                                        
#> 20124                                        
#> 20125                                        
#> 20126                                        
#>                                               homoSapiens
#>                                                    <list>
#> 1     ENSG00000105647,ENSG00000117461,ENSG00000145675,...
#> 2     ENSG00000021488,ENSG00000092068,ENSG00000103064,...
#> 3     ENSG00000021488,ENSG00000092068,ENSG00000103064,...
#> 4     ENSG00000003989,ENSG00000013293,ENSG00000021488,...
#> 5     ENSG00000003989,ENSG00000013293,ENSG00000021488,...
#> ...                                                   ...
#> 20122                                                    
#> 20123                                                    
#> 20124                                                    
#> 20125                                                    
#> 20126                                                    
#>                                                        musMusculus
#>                                                             <list>
#> 1     ENSMUSG00000028698,ENSMUSG00000031834,ENSMUSG00000041417,...
#> 2     ENSMUSG00000000958,ENSMUSG00000020600,ENSMUSG00000022180,...
#> 3     ENSMUSG00000000958,ENSMUSG00000020600,ENSMUSG00000022180,...
#> 4     ENSMUSG00000000958,ENSMUSG00000020600,ENSMUSG00000022180,...
#> 5     ENSMUSG00000000958,ENSMUSG00000020600,ENSMUSG00000022180,...
#> ...                                                            ...
#> 20122                                                             
#> 20123                                                             
#> 20124                                                             
#> 20125                                                             
#> 20126