Peptides

Peptides

peptides(
  version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam()
)

Arguments

version
character(1) or NULL. WormBase version. If NULL (recommended), defaults to current production release version available on the WormBase website. Legacy versions can be specified as a character string (e.g. "WS262").

BPPARAM
bpparamClass. BiocParallel parameter to specify the desired processor configuration.
We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

Value

SplitDataFrameList. Split by geneID column.

Note

Updated 2020-01-26.

Examples

## WormBase FTP server must be accessible.
tryCatch(
    expr = peptides(),
    error = function(e) e
)

#> Transmitting files from <ftp.wormbase.org>.

#> ℹ Files matching pattern: c_elegans.PRJNA13758.WS277.wormpep_package.tar.gz

#> → Downloading c_elegans.PRJNA13758.WS277.wormpep_package.tar.gz.

#> → Importing wormpep.table277 at /private/var/folders/l1/8y8sjzmn15v49jgrqglghcfr0000gn/T/RtmpRU0dkX using readr::`read_lines()`.

#> → Processing peptides.

#> SplitDataFrameList of length 19998
#> $WBGene00007063
#> DataFrame with 2 rows and 8 columns
#>           geneID  wormpep    status    uniprot      insdc sequence    locus
#>         <factor> <factor>  <factor>   <factor>   <factor> <factor> <factor>
#> 1 WBGene00007063  CE32090 Confirmed A4F336     CCD61130.1  2L52.1a       NA
#> 2 WBGene00007063  CE50569 Confirmed A0A0K3AWR5 CTQ86426.1  2L52.1b       NA
#>    product
#>   <factor>
#> 1       NA
#> 2       NA
#> 
#> $WBGene00007064
#> DataFrame with 3 rows and 8 columns
#>           geneID  wormpep    status    uniprot      insdc sequence    locus
#>         <factor> <factor>  <factor>   <factor>   <factor> <factor> <factor>
#> 1 WBGene00007064  CE32785 Confirmed A4F337     CCD61138.1 2RSSE.1a    rga-9
#> 2 WBGene00007064  CE48524 Confirmed U4PAK0     CDH92922.1 2RSSE.1b    rga-9
#> 3 WBGene00007064  CE52653 Confirmed A0A2X0T1Z3 SPS41576.1 2RSSE.1c    rga-9
#>                             product
#>                            <factor>
#> 1 Rho-GAP domain-containing protein
#> 2 Rho-GAP domain-containing protein
#> 3 Rho-GAP domain-containing protein
#> 
#> $WBGene00007065
#> DataFrame with 2 rows and 8 columns
#>           geneID  wormpep    status  uniprot      insdc sequence    locus
#>         <factor> <factor>  <factor> <factor>   <factor> <factor> <factor>
#> 1 WBGene00007065  CE24758 Confirmed   G5EFG7 CAA21777.2   3R5.1a    pot-3
#> 2 WBGene00007065  CE47782 Confirmed   J7SA53 CCM09397.1   3R5.1b    pot-3
#>                            product
#>                           <factor>
#> 1 POT1PC domain-containing protein
#> 2 POT1PC domain-containing protein
#> 
#> ...
#> <19995 more elements>