RNAi phenotypes

RNAi phenotypes

rnaiPhenotypes(
  version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam()
)

Arguments

version
character(1) or NULL. WormBase version. If NULL (recommended), defaults to current production release version available on the WormBase website. Legacy versions can be specified as a character string (e.g. "WS262").

BPPARAM
bpparamClass. BiocParallel parameter to specify the desired processor configuration.
We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

Value

DataFrame.

Note

Updated 2019-08-28.

Examples

## WormBase FTP server must be accessible.
tryCatch(
    expr = rnaiPhenotypes(),
    error = function(e) e
)

#> Transmitting files from <ftp.wormbase.org>.

#> ℹ Files matching pattern: rnai_phenotypes_quick.WS277.wb

#> → Downloading rnai_phenotypes_quick.WS277.wb.gz.

#> → Importing rnai_phenotypes_quick.WS277.wb.gz at /private/var/folders/l1/8y8sjzmn15v49jgrqglghcfr0000gn/T/RtmpRU0dkX using vroom::`vroom()`.

#> DataFrame with 7914 rows and 2 columns
#>              geneID
#>         <character>
#> 1    WBGene00000001
#> 2    WBGene00000002
#> 3    WBGene00000003
#> 4    WBGene00000004
#> 5    WBGene00000005
#> ...             ...
#> 7910 WBGene00305159
#> 7911 WBGene00305173
#> 7912 WBGene00305997
#> 7913 WBGene00305999
#> 7914 WBGene00306066
#>                                                                                           rnaiPhenotypes
#>                                                                                                   <list>
#> 1                                                          dauer induction variant,hermaphrodite sterile
#> 2                                                              chemical resistant,metal response variant
#> 3                                                              chemical resistant,metal response variant
#> 4                                        chemical resistant,fat content increased,metal response variant
#> 5    gonad development variant,male gonad development variant,male somatic gonad development variant,...
#> ...                                                                                                  ...
#> 7910                                                                                 fat content reduced
#> 7911             embryonic lethal,spontaneous mutation rate increased,transgene expression increased,...
#> 7912                                 chromosome segregation variant,embryonic lethal,meiosis variant,...
#> 7913                                                                               fat content increased
#> 7914                                                                                 fat content reduced