Best BLASTP Hits

Best BLASTP Hits

blastp(version = NULL, dir = ".")

Arguments

version
WormBase version. If NULL, defaults to current production release version available on the WormBase website. Legacy versions can be specified as a string (e.g. "WS262").

dir
Output directory.

Value

tbl_df grouped by wormpep column.

Examples

x <- blastp()

#> Files in directory:
#>  [1] "c_elegans.PRJNA13758.WS267.CDS_transcripts.fa.gz"        
#>  [2] "c_elegans.PRJNA13758.WS267.annotations.gff2.gz"          
#>  [3] "c_elegans.PRJNA13758.WS267.annotations.gff3.gz"          
#>  [4] "c_elegans.PRJNA13758.WS267.assembly.agp.gz"              
#>  [5] "c_elegans.PRJNA13758.WS267.best_blastp_hits.txt.gz"      
#>  [6] "c_elegans.PRJNA13758.WS267.canonical_geneset.gtf.gz"     
#>  [7] "c_elegans.PRJNA13758.WS267.ests.fa.gz"                   
#>  [8] "c_elegans.PRJNA13758.WS267.genomic.fa.gz"                
#>  [9] "c_elegans.PRJNA13758.WS267.genomic_masked.fa.gz"         
#> [10] "c_elegans.PRJNA13758.WS267.genomic_softmasked.fa.gz"     
#> [11] "c_elegans.PRJNA13758.WS267.intergenic_sequences.fa.gz"   
#> [12] "c_elegans.PRJNA13758.WS267.mRNA_transcripts.fa.gz"       
#> [13] "c_elegans.PRJNA13758.WS267.ncRNA_transcripts.fa.gz"      
#> [14] "c_elegans.PRJNA13758.WS267.protein.fa.gz"                
#> [15] "c_elegans.PRJNA13758.WS267.pseudogenic_transcripts.fa.gz"
#> [16] "c_elegans.PRJNA13758.WS267.transposon_transcripts.fa.gz" 
#> [17] "c_elegans.PRJNA13758.WS267.transposons.fa.gz"            
#> [18] "c_elegans.PRJNA13758.WS267.wormpep_package.tar.gz"       

#> Files matching pattern:
#> [1] "c_elegans.PRJNA13758.WS267.best_blastp_hits.txt.gz"

#> Downloading c_elegans.PRJNA13758.WS267.best_blastp_hits.txt.gz.

#> Parsed with column specification:
#> cols(
#>   .default = col_character(),
#>   X3 = col_double(),
#>   X5 = col_double(),
#>   X7 = col_double(),
#>   X9 = col_double(),
#>   X11 = col_double(),
#>   X13 = col_double(),
#>   X15 = col_double(),
#>   X17 = col_double(),
#>   X19 = col_double(),
#>   X21 = col_double(),
#>   X23 = col_double(),
#>   X25 = col_double(),
#>   X27 = col_double(),
#>   X29 = col_double()
#> )

#> See spec(...) for full column specifications.

glimpse(x)

#> Observations: 14,426
#> Variables: 3
#> $ wormpep <chr> "CE00004", "CE00015", "CE00034", "CE00063", "CE00072", "CE0...
#> $ peptide <chr> "ENSP00000308897", "ENSP00000444192", "ENSP00000384886", "E...
#> $ eValue  <dbl> 1e-46, 0e+00, 4e-19, 8e-08, 1e-20, 9e-66, 7e-07, 2e-25, 9e-...