Gene functional descriptions

Gene functional descriptions

description(
  version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam()
)

Arguments

version
character(1) or NULL. WormBase version. If NULL (recommended), defaults to current production release version available on the WormBase website. Legacy versions can be specified as a character string (e.g. "WS262").

BPPARAM
bpparamClass. BiocParallel parameter to specify the desired processor configuration.
We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

Value

DataFrame.

Note

As of WS269 release, some non-N2 gene IDs are included in the flat files available on the WormBase FTP server. These annotations are removed from the return here, using grep matching to return only WBGene entries.

This file is malformed on the WormBase FTP server for WS270 and WS271 releases.

Updated 2020-01-03.

Examples

## WormBase FTP server must be accessible.
tryCatch(
    expr = description(),
    error = function(e) e
)

#> Transmitting files from 'ftp.wormbase.org'.

#> Files matching pattern:
#> c_elegans.PRJNA13758.WS274.functional_descriptions.txt.gz

#> Downloading c_elegans.PRJNA13758.WS274.functional_descriptions.txt.gz.

#> Decompressing 'c_elegans.PRJNA13758.WS274.functional_descriptions.txt.gz' in 'tempdir()'.

#> Importing 'c_elegans.PRJNA13758.WS274.functional_descriptions.txt.gz' using 'base::readLines()'.

#> Processing functional descriptions.

#> DataFrame with 51300 rows and 4 columns
#>               geneID
#>             <factor>
#> 1     WBGene00000001
#> 2     WBGene00000002
#> 3     WBGene00000003
#> 4     WBGene00000004
#> 5     WBGene00000005
#> ...              ...
#> 51296 WBGene00304981
#> 51297 WBGene00304985
#> 51298 WBGene00304992
#> 51299 WBGene00304996
#> 51300 WBGene00305018
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            conciseDescription
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      <factor>
#> 1                                                                                                                                      aap-1 encodes the C. elegans ortholog of the phosphoinositide 3-kinase (PI3K) p50/p55 adaptor/regulatory subunit; AAP-1 negatively regulates lifespan and dauer development, and likely functions as the sole adaptor subunit for the AGE-1/p110 PI3K catalytic subunit to which it binds in vitro; although AAP-1 potentiates insulin-like signaling, it is not absolutely required for insulin-like signaling under most conditions.
#> 2      aat-1 encodes an amino acid transporter catalytic subunit; when co-expressed in Xenopus oocytes with the ATG-2 glycoprotein subunit, AAT-1 is able to facilitate amino acid uptake and exchange, showing a relatively high affinity for small and some large neutral amino acids; in addition, AAT-1 is able to covalently associate with ATG-2 or ATG-1 to form heterodimers in the Xenopus expression system; when co-expressed with ATG-2, AAT-1 localizes to the cell surface of oocytes, but when expressed alone or with ATG-1, AAT-1 localizes intracellularly.
#> 3                                                                                                                                                                                                                                                                                                                                                                               aat-2 encodes a predicted amino acid transporter catalytic subunit; when co-expressed in Xenopus oocytes with a glycoprotein subunit, however, AAT-2 is not able to induce amino acid uptake.
#> 4     aat-3 encodes an amino acid transporter catalytic subunit; when co-expressed in Xenopus oocytes with the ATG-2 glycoprotein subunit, AAT-3 is able to facilitate amino acid uptake and exchange, showing a relatively high affinity for small and some large neutral amino acids; in addition, AAT-3 is able to covalently associate with ATG-2 or ATG-1 to form heterodimers in the Xenopus expression system; when co-expressed with ATG-2, AAT-3 localizes to the cell surface of oocytes, but when expressed alone, or with ATG-1, AAT-3 localizes intracellularly.
#> 5                                                                                                                                                      aat-4 encodes a predicted amino acid transporter catalytic subunit; unlike catalytic subunits in other organisms, however, AAT-4 does not contain the highly conserved cysteine residue known to facilitate covalent interaction with a glycoprotein subunit, suggesting that AAT-4 does not require this residue for heterodimer formation or, alternatively, does not require the glycoprotein subunit for function.
#> ...                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       ...
#> 51296                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          none available
#> 51297                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          none available
#> 51298                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          none available
#> 51299                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          none available
#> 51300                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          none available
#>                                                                                                                                                                                                                                                                                                                                                                                                   automatedDescription
#>                                                                                                                                                                                                                                                                                                                                                                                                               <factor>
#> 1     Is an ortholog of human PIK3R3 (phosphoinositide-3-kinase regulatory subunit 3). Exhibits protein kinase binding activity. Is involved in dauer larval development; determination of adult lifespan; and insulin receptor signaling pathway. Localizes to the phosphatidylinositol 3-kinase complex. Is expressed in intestine and neurons. Human ortholog(s) of this gene are implicated in agammaglobulinemia.
#> 2                                                                           Is an ortholog of human SLC7A8 (solute carrier family 7 member 8). Contributes to L-amino acid transmembrane transporter activity. Is involved in amino acid transmembrane transport. Localizes to the amino acid transport complex. Is expressed in tail. Human ortholog(s) of this gene are implicated in lysinuric protein intolerance.
#> 3                                                                                                                                                                                             Is an ortholog of human SLC7A7 (solute carrier family 7 member 7). Is predicted to have L-amino acid transmembrane transporter activity. Human ortholog(s) of this gene are implicated in lysinuric protein intolerance.
#> 4                                                   Is an ortholog of human SLC7A5 (solute carrier family 7 member 5) and SLC7A8 (solute carrier family 7 member 8). Contributes to L-amino acid transmembrane transporter activity. Is involved in amino acid transmembrane transport. Localizes to the amino acid transport complex. Human ortholog(s) of this gene are implicated in lysinuric protein intolerance.
#> 5                                                                                                                                                                                                                               Is an ortholog of human SLC7A6 (solute carrier family 7 member 6) and SLC7A7 (solute carrier family 7 member 7). Is predicted to have L-amino acid transmembrane transporter activity.
#> ...                                                                                                                                                                                                                                                                                                                                                                                                                ...
#> 51296                                                                                                                                                                                                                                                                                                                                                                                                   none available
#> 51297                                                                                                                                                                                                                                                                                                                                                                                                   none available
#> 51298                                                                                                                                                                                                                                                                                                                                                                                                   none available
#> 51299                                                                                                                                                                                                                                                                                                                                                                                                   none available
#> 51300                                                                                                                                                                                                                                                                                                                                                                                                   none available
#>                          geneClassDescription
#>                                      <factor>
#> 1     phosphoinositide kinase AdAPter subunit
#> 2                      Amino Acid Transporter
#> 3                      Amino Acid Transporter
#> 4                      Amino Acid Transporter
#> 5                      Amino Acid Transporter
#> ...                                       ...
#> 51296                               not known
#> 51297                               not known
#> 51298                               not known
#> 51299                               not known
#> 51300                               not known