Peptides

Peptides

peptides(version = getOption(x = "wormbase.version", default = NULL),
  BPPARAM = BiocParallel::bpparam())

Arguments

version
character(1) or NULL. WormBase version. If NULL (recommended), defaults to current production release version available on the WormBase website. Legacy versions can be specified as a character string (e.g. "WS262").

Value

SplitDataFrameList. Split by geneID column.

Note

Updated 2019-08-28.

Examples

## WormBase FTP server must be accessible.
tryCatch(
    expr = peptides(),
    error = function(e) e
)

#> Transmitting files from 'ftp.wormbase.org'.

#> Files matching pattern:
#> c_elegans.PRJNA13758.WS271.wormpep_package.tar.gz

#> Downloading c_elegans.PRJNA13758.WS271.wormpep_package.tar.gz.

#> Importing 'wormpep.table271' using 'base::readLines()'.

#> Processing peptides.

#> SplitDataFrameList of length 20184
#> $WBGene00007063
#> DataFrame with 2 rows and 8 columns
#>           geneID  wormpep    status    uniprot      insdc sequence    locus
#>         <factor> <factor>  <factor>   <factor>   <factor> <factor> <factor>
#> 1 WBGene00007063  CE32090 Confirmed     A4F336 CCD61130.1  2L52.1a       NA
#> 2 WBGene00007063  CE50569 Confirmed A0A0K3AWR5 CTQ86426.1  2L52.1b       NA
#>    product
#>   <factor>
#> 1       NA
#> 2       NA
#> 
#> $WBGene00007064
#> DataFrame with 3 rows and 8 columns
#>           geneID  wormpep              status    uniprot      insdc sequence
#>         <factor> <factor>            <factor>   <factor>   <factor> <factor>
#> 1 WBGene00007064  CE32785 Partially_confirmed     A4F337 CCD61138.1 2RSSE.1a
#> 2 WBGene00007064  CE48524 Partially_confirmed     U4PAK0 CDH92922.1 2RSSE.1b
#> 3 WBGene00007064  CE52653 Partially_confirmed A0A2X0T1Z3 SPS41576.1 2RSSE.1c
#>      locus  product
#>   <factor> <factor>
#> 1    rga-9       NA
#> 2    rga-9       NA
#> 3    rga-9       NA
#> 
#> $WBGene00007065
#> DataFrame with 2 rows and 8 columns
#>           geneID  wormpep              status  uniprot      insdc sequence
#>         <factor> <factor>            <factor> <factor>   <factor> <factor>
#> 1 WBGene00007065  CE24758 Partially_confirmed   G5EFG7 CAA21777.2   3R5.1a
#> 2 WBGene00007065  CE47782           Confirmed   J7SA53 CCM09397.1   3R5.1b
#>      locus  product
#>   <factor> <factor>
#> 1    pot-3       NA
#> 2    pot-3       NA
#> 
#> ...
#> <20181 more elements>