RNAi Phenotypes

RNAi Phenotypes

rnaiPhenotypes(version = NULL, dir = ".")

Arguments

version
WormBase version. If NULL, defaults to current production release version available on the WormBase website. Legacy versions can be specified as a string (e.g. "WS262").

dir
Output directory.

Value

tbl_df.

Examples

invisible(capture.output(
    x <- rnaiPhenotypes()
))

#> Files in directory:
#>  [1] "anatomy_association.WS267.wb"         
#>  [2] "anatomy_ontology.WS267.obo"           
#>  [3] "development_association.WS267.wb"     
#>  [4] "development_ontology.WS267.obo"       
#>  [5] "disease_association.WS267.daf.txt"    
#>  [6] "disease_association.WS267.wb"         
#>  [7] "disease_ontology.WS267.obo"           
#>  [8] "gene_association.WS267.wb"            
#>  [9] "gene_association.WS267.wb.b_malayi"   
#> [10] "gene_association.WS267.wb.c_brenneri" 
#> [11] "gene_association.WS267.wb.c_briggsae" 
#> [12] "gene_association.WS267.wb.c_elegans"  
#> [13] "gene_association.WS267.wb.c_japonica" 
#> [14] "gene_association.WS267.wb.c_remanei"  
#> [15] "gene_association.WS267.wb.o_volvulus" 
#> [16] "gene_association.WS267.wb.p_pacificus"
#> [17] "gene_association.WS267.wb.s_ratti"    
#> [18] "gene_association.WS267.wb.t_muris"    
#> [19] "gene_ontology.WS267.obo"              
#> [20] "phenotype_association.WS267.wb"       
#> [21] "phenotype_ontology.WS267.obo"         
#> [22] "rnai_phenotypes.WS267.wb"             
#> [23] "rnai_phenotypes_quick.WS267.wb"       

#> Files matching pattern:
#> [1] "rnai_phenotypes_quick.WS267.wb"

#> Downloading rnai_phenotypes_quick.WS267.wb.gz.

#> Parsed with column specification:
#> cols(
#>   geneID = col_character(),
#>   sequence = col_character(),
#>   rnaiPhenotypes = col_character()
#> )

glimpse(x)

#> Observations: 8,116
#> Variables: 2
#> $ geneID         <chr> "WBGene00001908", "WBGene00019433", "WBGene00002090"...
#> $ rnaiPhenotypes <list> [<"accumulated cell corpses", "anaphase bridging", ...